Rotations are on the same device.

+++ odak/learn/raytracing/primitives.py
@@ -22,7 +22,7 @@ def define_plane(point, angles=[0., 0., 0.]):
         [10., 10., 0.],
         [0., 10., 0.],
         [0.,  0., 0.]
-    ], dtype=torch.float).to(point.device)
+    ]).to(point.device)
     for i in range(0, plane.shape[0]):
         plane[i], _, _, _ = rotate_point(plane[i], angles = angles)
         plane[i] = plane[i] + torch.tensor(point)
